==> AI/DeepLearning/GNN_Reactome/2.Pytorch_Graph_GNN.ipynb <==
import torch
import numpy as np
import pandas as pd
import copy
import torch.nn.functional as F
import scipy.sparse as sp
from torch_geometric.nn import GCNConv,GATConv,SAGEConv
from torch_geometric.datasets import Planetoid
from torch.nn import Linear

# 把标签转换成onehot
def encode_onehot(labels):                                   
    classes = set(labels)
    classes_dict = {c: np.identity(len(classes))[i, :] for i, c in enumerate(classes)}
    labels_onehot = np.array(list(map(classes_dict.get, labels)), dtype=np.int64)
    return labels_onehot

def normalize(mx):                                          # 归一化
    rowsum = np.array(mx.sum(1))
    r_inv = np.power(rowsum, -1).flatten() ## 幂
    r_inv[np.isinf(r_inv)] = 0.
    r_mat_inv = sp.diags(r_inv) # 对角化
    mx = r_mat_inv.dot(mx) # 乘积
    #mx = mx * 10000
    return mx

## 获得稳定的Tissue标签对应的数字
GTEx_exp_full = pd.read_csv("/public/home/liujunwu/workdir/scripts/GNN_Reactome/GTEx_exp/GTEx_exp_down3000.csv",sep=',',header=0)


print (GTEx_exp_full.iloc[0:3,0:3])
sample_full_labels = sorted(list(set(GTEx_exp_full["SMTS"].tolist())))
sample_full_labels_dict = dict(zip(sample_full_labels,range(len(sample_full_labels))))

print (list(set(GTEx_exp_full["Tissue"])))
sample_full_labels = sorted(list(set(GTEx_exp_full["Tissue"].tolist())))
sample_full_labels_dict = dict(zip(sample_full_labels,range(len(sample_full_labels))))
print (sample_full_labels_dict)

# 读取GTEx 表达谱文件
GTEx_exp = pd.read_csv("/public/home/liujunwu/workdir/scripts/GNN_Reactome/GTEx_exp/GTEx_exp.csv.test",sep=',',header=0)
sample_labels = np.where(encode_onehot(GTEx_exp["Tissue"]))[1]
#sample_labels = encode_onehot(GTEx_exp["Tissue"])
#print (sample_labels)
# 存储标签和真实tissue关系
sample_labels_dict = dict(zip(sample_labels,GTEx_exp["Tissue"]))
print (sample_labels_dict)
print (sample_labels_dict[7])
list_labels = []
for x in list(sample_labels):
    list_labels.append([x])
sample_labels = torch.tensor(list_labels,dtype=torch.int64)
#print (sample_labels)
print (sample_labels.dtype)
print (sample_labels[0])
print (sample_labels[0].shape)
print (len(sample_labels_dict.keys()))

reaction_nodes = pd.read_csv("/public/home/liujunwu/workdir/scripts/GNN_Reactome/reaction_file/reactome_reaction.nodes.txt",sep='\t',header=None)
print (reaction_nodes.shape)
reaction_nodes.columns = ["Reaction","Type","Gene"]
reaction_gene_list = []
for index,row in reaction_nodes.iterrows():
    if row[2] != "" :
        reaction_gene_list.extend(str(row[2]).split(","))
common_gene_list = sorted(list(set(reaction_gene_list) & set(GTEx_exp.columns)))
common_gene_list_dict = dict(zip(sorted(common_gene_list),range(len(common_gene_list))))
#存储节点和index关系
reaction_nodes_dict = dict(zip(reaction_nodes.index,reaction_nodes["Reaction"]))
reaction_nodes_rev_dict = dict(zip(reaction_nodes["Reaction"],reaction_nodes.index))
## 初始化0矩阵
zero_matrix = pd.DataFrame(np.zeros((reaction_nodes.shape[0],len(common_gene_list))),columns = common_gene_list) 
print (zero_matrix.shape)
print (reaction_nodes.iloc[0:3,0:3])
print (zero_matrix.iloc[0:3,0:3])

# 精简 GTEx_exp 的维度
GTEx_exp_match = GTEx_exp.loc[:,list(common_gene_list)]
print (GTEx_exp_match.shape)

print ("SLC66A1" in GTEx_exp_match.columns)

edges_file = "/public/home/liujunwu/workdir/scripts/GNN_Reactome/reaction_file/reactome_reaction.edges.uniq.txt"
edges_pd = pd.read_csv(edges_file,header=None,sep='\t')
#edges_pd.columns = ["Edge1","Edge2","Type","TopLevel"]
edges_pd.columns = ["Edge1","Edge2"]
edges_pd["Edge1_index"] = edges_pd["Edge1"].apply(lambda x : reaction_nodes_rev_dict[x])
edges_pd["Edge2_index"] = edges_pd["Edge2"].apply(lambda x : reaction_nodes_rev_dict[x])
#edges_trans.loc[df["Edge2"]] = edges_trans["Edge2"].apply(lambda x : reaction_nodes_dict[x])
print (edges_pd[["Edge1_index","Edge2_index"]])
edges_trans = edges_pd[["Edge1_index","Edge2_index"]]
#print (edges_trans[0:3,0:1])
#edges_trans = torch.tensor(np.ndarray(edges_trans).todense(), dtype=torch.int64)
#print (edges_trans)
edges = []
for i in range(edges_trans.shape[0]):
    edges.append([edges_trans.at[i,"Edge1_index"],edges_trans.at[i,"Edge2_index"]])
edges = torch.tensor(edges,dtype = torch.int64).T
print (edges)

from sys import getsizeof as getsize

not_single_nodes = reaction_nodes.query('Type == "ProteinReaction" | Type == "ComplexReaction"')
print (single_nodes.iloc[0:3,0:3])


import datetime
from torch_geometric.data import Data
print (zero_matrix.shape)
print (GTEx_exp_match.shape)
gene_index_dict = dict(zip(zero_matrix.columns.to_list(),range(zero_matrix.shape[1])))
print(datetime.datetime.now().strftime('%Y-%m-%d  %H:%M:%S'))
GTEx_exp_ndarry =  GTEx_exp_match.values
## 每个样本的边根据入度生成小图，然后通过batch组成大图
dataset = []
for sample_index in range(GTEx_exp_match.shape[0]):
    print (sample_index)
    sample_node_feature = copy.deepcopy(zero_matrix.values)
    for reaction_index in range(reaction_nodes.shape[0]):
        if (reaction_nodes.at[reaction_index,"Type"] not in ["ProteinReaction","ComplexReaction"]):continue # 跳过非蛋白对应的reaction
        current_reaction_gene_index = [common_gene_list_dict[x] for x in reaction_nodes_dict[reaction_index] if x in common_gene_list]
        sample_node_feature[[reaction_index],[current_reaction_gene_index]] = GTEx_exp_ndarry[[sample_index],[current_reaction_gene_index]]
    sample_node_feature = Data(x=torch.tensor(sample_node_feature,dtype = torch.float32),y=sample_labels[sample_index],edge_index = edges)
    dataset.append(sample_node_feature)
    print(datetime.datetime.now().strftime('%Y-%m-%d  %H:%M:%S'))
    #break
print ("store all sample node features done")
print (type(dataset))
print(datetime.datetime.now().strftime('%Y-%m-%d  %H:%M:%S'))

print (getsize(dataset))

print (sp.csc_matrix(sample_node_feature))

print (dataset[1])

print (type(dataset[0].num_))

from torch_geometric.nn import GraphConv

class GNN(torch.nn.Module):
    def __init__(self,num_node_features, hidden_channels,num_classes):
        super(GNN, self).__init__()
        torch.manual_seed(12345)
        self.conv1 = GCNConv(num_node_features, hidden_channels)
        self.conv2 = GraphConv(hidden_channels, hidden_channels)
        self.conv3 = GraphConv(hidden_channels, hidden_channels)
        #print (hidden_channels.dtype)
        #print (num_classes.dtype)
        self.lin = Linear(hidden_channels,num_classes)

    def forward(self, x, edge_index,batch):
        x = self.conv1(x, edge_index)
        x = x.relu()
        x = self.conv2(x, edge_index)
        x = x.relu()
        x = self.conv3(x, edge_index)
        #return F.log_softmax(x, dim=1)
        x = global_mean_pool(x, batch)
        x = F.dropout(x, p=0.5, training=self.training)
        #print (x.dtype)
        x = self.lin(x)
        return x


model = GNN(num_node_features=7979,hidden_channels=64,num_classes=len(sample_labels_dict.keys()))
print(model)
#print (num_classes.dtype)

print (dataset[0].x.dtype)
print (dataset[0].y.dtype)
print (dataset[0].edge_index.dtype)

del GTEx_exp
del zero_matrix


optimizer = torch.optim.Adam(model.parameters(), lr=0.01)
criterion = torch.nn.CrossEntropyLoss()
torch.set_num_threads(48)
from torch_geometric.nn import global_mean_pool
from torch_geometric.loader import DataLoader
train_dataset = dataset[0:40]
train_loader = DataLoader(train_dataset, batch_size=1,shuffle=True)

def train():
    model.train()
    for data in dataset:
        #print (data.y)
        optimizer.zero_grad()
        out = model(data.x, data.edge_index,data.batch)
        loss = criterion(out, data.y)
        loss.backward()
        optimizer.step()
def test(loader):
    model.eval()
    correct = 0
    for data in loader:                            # 批遍历测试集数据集。
        out = model(data.x, data.edge_index, data.batch) # 一次前向传播
        pred = out.argmax(dim=1)                         # 使用概率最高的类别
        correct += int((pred == data.y).sum())           # 检查真实标签
    return correct / len(loader.dataset)

model.train()
for epoch in range(1, 201):
    train()
    train_acc = test(train_loader)
    print(f'Epoch: {epoch:03d}, Train Acc: {train_acc:.4f}')

torch.save(dataset,"/public/home/liujunwu/workdir/scripts/GNN_Reactome/GTEx_exp/Torch_split/test.ptch")

print (dataset[0])

import torch
from torch_geometric.nn import MessagePassing
from torch_geometric.utils import add_self_loops, degree
 
class GCNConv(MessagePassing):
    def __init__(self, in_channels, out_channels):
        super().__init__(aggr='add') 
        # "Add" aggregation (Step 5).
        #GCN类从MessagePssing中继承得到的聚合方式：“add”
        self.lin = torch.nn.Linear(in_channels, out_channels)
 
    def forward(self, x, edge_index):
        # x has shape [N, in_channels] ——N个点，每个点in_channels维属性
        # edge_index has shape [2, E]——E条边，每条边有出边和入边
 
        # Step 1: Add self-loops to the adjacency matrix.
        edge_index, _ = add_self_loops(edge_index, num_nodes=x.size(0))
        #添加自环
 
        # Step 2: Linearly transform node feature matrix.
        x = self.lin(x)
        #对X进行线性变化
 
        # Step 3: Compute normalization.
        row, col = edge_index
        #出边和入边
        deg = degree(col, x.size(0), dtype=x.dtype)
        #各个点的入度（无向图，所以入读和出度相同）
        deg_inv_sqrt = deg.pow(-0.5)
        deg_inv_sqrt[deg_inv_sqrt == float('inf')] = 0
        norm = deg_inv_sqrt[row] * deg_inv_sqrt[col]
        #1/sqrt(di) *1/sqrt(dj)
 
        # Step 4-5: Start propagating messages.
        return self.propagate(edge_index, x=x, norm=norm)
        #进行propagate
        #propagate的内部会调用message(),aggregate()和update()
        #作为消息传播的附加参数，我们传递节点嵌入x和标准化系数norm。    
 
    def message(self, x_j, norm):
        # x_j has shape [E, out_channels]
        #我们需要对相邻节点特征x_j进行norm标准化
        #这里x_j为一个张量，其中包含每条边的源节点特征，即每个节点的邻居。
 
        # Step 4: Normalize node features.
        return norm.view(-1, 1) * x_j
        #1/sqrt(di) *1/sqrt(dj) *X_j

conv = GCNConv(7979, 64)
x = conv(dataset[0].x,edges )
print (edges.shape)
print (dataset[0].x.shape)
print (x.shape)

from torch_geometric.data import Dataset, download_url


class MyOwnDataset(Dataset):
    def __init__(self, root, transform=None, pre_transform=None, pre_filter=None):
        super().__init__(root, transform, pre_transform, pre_filter)

    @property
    def raw_file_names(self):
        return ['some_file_1', 'some_file_2', ...]

    @property
    def processed_file_names(self):
        return ['data_1.pt', 'data_2.pt', ...]

    #def download(self):
        # Download to `self.raw_dir`.
     #   path = download_url(url, self.raw_dir)
     #   ...

    def process(self):
        idx = 0
        for raw_path in self.raw_paths:
            # Read data from `raw_path`.
            data = Data(...)

            if self.pre_filter is not None and not self.pre_filter(data):
                continue

            if self.pre_transform is not None:
                data = self.pre_transform(data)

            torch.save(data, osp.join(self.processed_dir, f'data_{idx}.pt'))
            idx += 1

    def len(self):
        return len(self.processed_file_names)

    def get(self, idx):
        data = torch.load(osp.join(self.processed_dir, f'data_{idx}.pt'))
        return data

import os
print (os.getcwd())
print(model)
print (globals().keys())

#删除clear_env.py
def __clear_env():
    for key in globals().keys():
        if not key.startswith("__"):# 排除系统内建函数
            globals().pop(key)
_clear_env()

from torch.utils.data import Dataset

class Mydataset(Dataset):
    def __init__(self,):
        self.datas=[1,2,(4,'a')]
    
    def __len__(self,):
        return len(self.datas)
    
    def __getitem__(self,index):
        
        return self.datas[index]

mydataset = Mydataset()

mydataset[0]

from torch_geometric.data import Data
from torch_geometric.data import DataLoader
print (len(dataset))
print (os.getcwd())
loader = DataLoader(dataset[0:10],batch_size=32,shuffle=True)

print (len(loader))

sample_node_feature = Data(x=torch.tensor([sample_node_feature,dtype = torch.float32),y=sample_labels[sample_index],edge_index = edges)

from torch_geometric.data import Data
from torch_geometric.data.batch import Batch
edge_index_s = torch.tensor([
    [0, 0, 0, 0],
    [1, 2, 3, 4],
])
x_s = torch.randn(5, 8864)  # 5 nodes.
edge_index_t = torch.tensor([
    [0, 0, 0],
    [1, 2, 3],
])
x_t = torch.randn(4, 8864)  # 4 nodes.

edge_index_3 = torch.tensor([[0, 1, 1, 2],
                           [1, 0, 2, 1]], dtype=torch.long)
x_3 = torch.randn(4, 8864)

data1= Data(x=x_s,edge_index=edge_index_s)
data2= Data(x=x_t,edge_index=edge_index_t)
data3= Data(x=x_3,edge_index=edge_index_3)
data4= Data(x=x_s,edge_index=edge_index_3)
#上面是构建3张Data图对象
# * `Batch(Data)` in case `Data` objects are batched together
#* `Batch(HeteroData)` in case `HeteroData` objects are batched together

data_list = [data1, data2,data3,data4]
loader = Batch.from_data_list(data_list)#调用该函数data_list里的data1、data2、data3 三张图形成一张大图，也就是batch
loader2 = Batch.from_data_list([data1,data3])

from sys import getsizeof
print (loader)
print (loader2)
print (getsizeof(loader))
print (getsizeof(data1))
getsize(data2)
getsize(data3)
getsize(data4)
