==> AI/DeepLearning/GNN_Reactome/3.Reactome_PathwayGSEA_GNN.ipynb <==
import torch
import numpy as np
import pandas as pd
import copy
import torch.nn.functional as F
import scipy.sparse as sp
from torch_geometric.nn import GCNConv,GATConv,SAGEConv
from torch_geometric.datasets import Planetoid
from torch.nn import Linear
from sklearn.preprocessing import StandardScaler
from sklearn import preprocessing

# 对样本进行归一化，而不是特征
def Scale(X):
    raw = X
    scaler = StandardScaler()
    scaler.fit(X.T)
    X = pd.DataFrame(scaler.transform(X.T)).T
    X.columns = raw.columns
    X.index = raw.index
    return X

def ProssGSEA(csv):
    dataframe = pd.read_csv(csv,sep=",")
    dataframe = dataframe.pivot(index = "Name",columns=["Term"],values=["NES"])
    dataframe.columns = dataframe.columns.droplevel(0)
    dataframe.index.rename(None, inplace=True)
    dataframe = dataframe.dropna(axis=1, how='all') ## 去除所有都是NaN的列
    return dataframe

ReactomePathways_file = "/public/ref/Msigdb/Reactome/useInfo/Human.ReactomePathways.list"
ReactomePathways_label = pd.read_csv(ReactomePathways_file,header=None,sep='\t')
#print (ReactomePathways_label.iloc[0:1,0:1])
ReactomePathways_label.columns = ["Name","ID"]
ReactomePathways_label_dict = dict(zip(ReactomePathways_label["Name"],ReactomePathways_label["ID"])) ## Name : R-HSA*
ReactomePathways_label_rev_dict = dict(zip(ReactomePathways_label["ID"],ReactomePathways_label["Name"])) # R-HSA*: Name
print (ReactomePathways_label_rev_dict["R-HSA-9694493"])
print (ReactomePathways_label_dict["Maturation of protein E_9694493"])
print (ReactomePathways_label_dict['Hereditary fructose intolerance'])


#GSEA_file = "/public/home/liujunwu/workdir/scripts/GNN_Reactome/GTEx_exp/GSEA/GSEA_5tissue50.csv"
GSEA_file = "/public/home/liujunwu/workdir/scripts/GNN_Reactome/GTEx_exp/GSEA/GSEA_all.csv"
#GSEA_score = pd.read_csv(GSEA_file,header=0,sep=',')
#print (GSEA_score.iloc[0:3,0:3])
#GSEA_score["Term"] = GSEA_score["Term"].map(ReactomePathways_label_dict)  ## 某一列替换为对应字典值
#GSEA_score = GSEA_score.pivot(index="Name",columns="Term",values="NES")
#GSEA_score.index.rename(None, inplace=True)
#GSEA_score = GSEA_score.dropna(axis=1, how='all')
GSEA_score =ProssGSEA(GSEA_file)
#GSEA_score = Scale(GSEA_score)
print (GSEA_score.shape)
print (GSEA_score.iloc[0:3,0:5])
#GSEA_score_ndarry =  GSEA_score.values
GSEA_score_ndarry = preprocessing.scale(GSEA_score.values,axis=1)
print (GSEA_score.values[[2],[3]])


print (GSEA_score.loc["GTEX-1117F-0426-SM-5EGHI"])

edges_file = "/public/home/liujunwu/workdir/scripts/GNN_Reactome/reaction_file/Human.pathway.Hierarchy.txt"
edges_pd = pd.read_csv(edges_file,header=None,sep='\t')
#edges_pd.columns = ["Edge1","Edge2","Type","TopLevel"]
edges_pd.columns = ["Edge1","Edge2"]
#print (edges_pd["Edge1"])
#print (type(list(edges_pd['Edge1'])))
print (edges_pd.shape)
edges_pd = edges_pd.loc[(edges_pd["Edge1"].isin(GSEA_score.columns)) &  edges_pd["Edge2"].isin(GSEA_score.columns)]
edges_pd.reset_index()
print (edges_pd.shape)
edges_list = sorted(pd.concat([edges_pd['Edge1'],edges_pd['Edge2']]).unique().tolist())
edges_dict = dict(zip(edges_list,range(len(edges_list))))
edges_dict_rev = dict(zip(range(len(edges_list)),edges_list))
print ("R-HSA-109582" in edges_list)
print (edges_pd.at[10,"Edge1"])
print (edges_dict[edges_pd.at[10,"Edge1"]])
#print (edges)
edges = list()
print (edges_pd.at[97,"Edge1"])
for i in range(edges_pd.shape[0]):
    #print (edges_pd.at[i,"Edge1"])
    #print (i)
    edges.append([edges_dict[edges_pd.iat[i,0]],edges_dict[edges_pd.iat[i,1]]])
    #print (edges)
edges = torch.tensor(edges,dtype = torch.int64).T
print (edges)
print (edges.shape)

print (ReactomePathways_label_rev_dict[edges_dict_rev[2]])
torch.save(edges,"/public/home/liujunwu/workdir/scripts/GNN_Reactome/GTEx_exp/pathway_edges.pt")

print (GSEA_score_ndarry[[2],[5]])

# 准备sample对应的tissue和对应的label 标签 ()
#label_file = "/public/home/liujunwu/workdir/scripts/GNN_Reactome/GTEx_exp/GSEA/GSEA_5tissue50.label.txt"
label_file = "/public/home/liujunwu/workdir/scripts/GNN_Reactome/GTEx_exp/GSEA/GTEx.sample_label.txt"
sample_labels = pd.read_csv(label_file,header=0,sep="\t")
sample_full_labels = sorted(list(set(sample_labels["SMTS"].tolist()))) ## 唯一值
sample_full_labels_dict = dict(zip(sample_full_labels,range(len(sample_full_labels)))) ## label对应的index
list_labels = []
print (sample_full_labels_dict)
sample_dict = dict(zip(sample_labels["SAMPID"],sample_labels["SMTS"]))
print (sample_full_labels_dict[sample_dict["GTEX-1GF9W-1326-SM-7P8PX"]])
for x in range(len(sample_full_labels)):
    list_labels.append([x])
list_labels_tensor = torch.tensor(list_labels,dtype=torch.int64)
print (list_labels_tensor[sample_full_labels_dict[sample_dict["GTEX-1GF9W-1326-SM-7P8PX"]]])

GSEA_score_ndarry[[2]]

#print (GSEA_score.loc["GTEX-117YW-2226-SM-5N9DB"])
print (GSEA_score.shape)


import datetime
from torch_geometric.data import Data
uniq_sample = sorted(list(set(GSEA_score.index)))
train_dataset = list()
print(datetime.datetime.now().strftime('%Y-%m-%d  %H:%M:%S'))
for sampleIndex,sampleName in enumerate(uniq_sample):
    #print (sampleName)
    sampleValue = GSEA_score_ndarry[sampleIndex]
    sampleValue = sampleValue.reshape(sampleValue.shape[0],1)
    #print (list_labels_tensor[sample_full_labels_dict[sample_dict[sampleName]]])
    sample_node_feature = Data(x=torch.tensor(sampleValue,dtype = torch.float32),y=list_labels_tensor[sample_full_labels_dict[sample_dict[sampleName]]],edge_index = edges)
    train_dataset.append(sample_node_feature)
    #break
print(datetime.datetime.now().strftime('%Y-%m-%d  %H:%M:%S'))
print (len(train_dataset))

print (train_dataset[0])

#a = GSEA_score.loc["GTEX-1GF9W-1326-SM-7P8PX"]
#a = torch.tensor(a,dtype = torch.float32)
print (train_dataset[0])
print (train_dataset[0].x)
print (train_dataset[0].y)
#print (type(a))

from torch.nn import Linear
from torch_geometric.nn import GraphConv

class GNN(torch.nn.Module):
    def __init__(self, num_node_features,hidden_channels,num_classes):
        super(GNN, self).__init__()
        torch.manual_seed(12345)
        self.conv1 = GraphConv(num_node_features, hidden_channels)
        self.conv2 = GraphConv(hidden_channels, hidden_channels)
        self.conv3 = GraphConv(hidden_channels, hidden_channels)
        self.lin = Linear(hidden_channels,num_classes)
        
    def forward(self, x, edge_index, batch):
        # 1. 获得节点嵌入
        x = self.conv1(x, edge_index)
        x = x.relu()
        x = self.conv2(x, edge_index)
        x = x.relu()
        x = self.conv3(x, edge_index)
        
        # 2. Readout layer
        x = global_mean_pool(x, batch)   # [batch_size, hidden_channels]
        
        # 3. 分类器
        x = F.dropout(x, p=0.5, training=self.training)
        x = self.lin(x)
        return x

#model = GNN(num_node_features=2582,hidden_channels=64,num_classes=len(sample_full_labels_dict.keys()))
model = GNN(num_node_features=1,hidden_channels=64,num_classes=31)
print (model)

from torch_geometric.nn import global_mean_pool
from torch_geometric.loader import DataLoader
optimizer = torch.optim.Adam(model.parameters(), lr=0.001)
criterion = torch.nn.CrossEntropyLoss()
torch.set_num_threads(48)
print (train_dataset[0])
train_dataset1 = train_dataset[0:17000]
train_loader = DataLoader(train_dataset1, batch_size=30,shuffle=True)
def train():
    model.train()
    for data in train_loader:
        #print (data.y)
        optimizer.zero_grad()
        out = model(data.x, data.edge_index,data.batch)
        loss = criterion(out, data.y)
        loss.backward()
        optimizer.step()
def test(loader):
    model.eval()
    correct = 0
    for data in loader:                            # 批遍历测试集数据集。
        out = model(data.x, data.edge_index, data.batch) # 一次前向传播
        pred = out.argmax(dim=1)   # 使用概率最高的类别
        #print (pred,data.y)
        correct += int((pred == data.y).sum())           # 检查真实标签
    return correct / len(loader.dataset)

model.train()
print(datetime.datetime.now().strftime('%Y-%m-%d  %H:%M:%S'))

for epoch in range(1, 301):
    train()
    print(datetime.datetime.now().strftime('%Y-%m-%d  %H:%M:%S'))
    train_acc = test(train_loader)
    print(f'Epoch: {epoch:03d}, Train Acc: {train_acc:.4f}')

print (train_loader[0].batch)

def test(loader):
    model.eval()
    correct = 0
    for data in loader:                            # 批遍历测试集数据集。
        out = model(data.x, data.edge_index, data.batch) # 一次前向传播
        pred = out.argmax(dim=1)   # 使用概率最高的类别
        print (pred,data.y)
        correct += int((pred == data.y).sum())           # 检查真实标签
    return correct / len(loader.dataset)

# 测试集
model.eval()
test_data = train_dataset[16000:]
print (len(test_data))
test_loader = DataLoader(test_data, batch_size=10,shuffle=True)
test_acc = test(test_loader)
#print (test_acc.)
print(f'Test Acc: {test_acc:.4f}')

model.eval()
#print (test_loader)
data = test_data[1]
#print (data)
out = model(data.x, data.edge_index,data.batch) # 一次前向传播
print (out)
pred = out.argmax(dim=1)
print (pred)

# 保存模型
saved_dict = {
 'model': model.state_dict(),
 'opt': optimizer.state_dict()
}
torch.save(saved_dict, '/public/home/liujunwu/workdir/scripts/GNN_Reactome/GTEx_exp/GTEx.nomalizeSample.axis1.pth.tar')

#TCGA_GSEA_score = TotalSample_GSEA.csv 
TCGA_GSEA_score = pd.read_csv("/public/home/liujunwu/workdir/scripts/GNN_Reactome/TCGA_exp/GSEA/TotalSample_GSEA.csv",header=0,sep=',')
TCGA_GSEA_score = TCGA_GSEA_score.pivot(index="Name",columns="Term",values="NES")
TCGA_GSEA_score.index.rename(None, inplace=True)
TCGA_GSEA_score = TCGA_GSEA_score.dropna(axis=1, how='all')
print (TCGA_GSEA_score.shape)
print (TCGA_GSEA_score.iloc[0:3,0:5])
TCGA_GSEA_score_ndarry = preprocessing.scale(TCGA_GSEA_score.values,axis=1)
print (TCGA_GSEA_score_ndarry[[2],[3]])

## 随机切割 数据
from torch.utils.data import random_split
TCGA_train_dataset, TCGA_test_dataset = random_split(
    dataset=tuning_dataset,
    lengths=[365, 365],
    generator=torch.Generator().manual_seed(0)
)
print (len(TCGA_train_dataset))
print (len(TCGA_test_dataset))
