==> AI/DeepLearning/GNN_Reactome/7.Bulk_Weighted_Edges.ipynb <==
import torch
import numpy as np
import pandas as pd
import copy
import torch.nn.functional as F
import scipy.sparse as sp
from torch_geometric.nn import GCNConv,GATConv,SAGEConv
from torch_geometric.datasets import Planetoid
from torch.nn import Linear
from sklearn.preprocessing import StandardScaler
from sklearn import preprocessing
import datetime
from torch_geometric.data import Data
from torch_geometric.loader import DataLoader
from torch_geometric.nn import global_mean_pool

#device = "cpu"
device = "cuda" if torch.cuda.is_available() else "cpu"
print("Using {} device".format(device))
torch.cuda.set_device(1)
torch.cuda.current_device()

device = cpu = torch.device('cpu')

#os.environ['CUDA_LAUNCH_BLOCKING'] = "1"
print (device)

from torch.nn import Linear
from torch_geometric.nn import GraphConv

class GNN(torch.nn.Module):
    def __init__(self, num_node_features,hidden_channels,num_classes):
        super(GNN, self).__init__()
        torch.manual_seed(12345)
        self.conv1 = GraphConv(num_node_features, hidden_channels)
        self.conv2 = GraphConv(hidden_channels, hidden_channels)
        self.conv3 = GraphConv(hidden_channels, hidden_channels)
        self.lin = Linear(hidden_channels,num_classes)
        
    def forward(self, x, edge_index, batch, edge_weight=None):
        # 1. 获得节点嵌入
        x = self.conv1(x, edge_index,edge_weight)
        x = x.relu()
        x = self.conv2(x, edge_index,edge_weight)
        x = x.relu()
        x = self.conv3(x, edge_index,edge_weight)
        
        # 2. Readout layer
        x = global_mean_pool(x, batch)   # [batch_size, hidden_channels]
        
        # 3. 分类器
        x = F.dropout(x,p=0.5,training=self.training)
        x = self.lin(x)
        return x

TORCH_USE_CUDA_DSA

## 读取reaction包含的基因
ReactionGeneFile = "/public/home/liujunwu/workdir/scripts/GNN_Reactome/reaction_file/reactome_reaction.ProteinReactionNodes.txt"
ReactionGene = pd.read_csv(ReactionGeneFile,header=None,sep='\t')
ReactionGene.columns = ["Reaction","Gene"]
ReactionGene_dict = dict(zip(ReactionGene["Reaction"],ReactionGene["Gene"])) ## Reaction 对应的Gene
print (ReactionGene_dict['R-HSA-1112666'])
del ReactionGene

edges_file = "/public/home/liujunwu/workdir/scripts/GNN_Reactome/reaction_file/reactome_reaction.edges.txt"
edges = pd.read_csv(edges_file,header=None,sep="\t")
edges.columns = ["edge1","edge2","edge1_type","edge2_type"] 
uniq_output_nodes = sorted(list(set(edges["edge2"].tolist()))) ## 取得唯一入度的点
uniq_nodes = sorted(list(set(edges["edge1"].tolist()) | set(edges["edge2"].tolist())))
print (len(uniq_nodes))
edges = torch.load("/public/home/liujunwu/workdir/scripts/GNN_Reactome/reaction_file/Reactome_reaction_edges.pt")
print (edges)

## 读取 IBD不同药物反应的bulk数据
metainfo = pd.read_csv("/public/data/scDT/SingleCellDownload/IBD/Treatment_Bulk/Bulk_Anti-TNF/GSE16879/GSE16879.metainfo",header=0,sep=',')
#labels = ["Control","BeforeT_Yes","BeforeT_No"]
labels = ["BeforeT_Yes","BeforeT_No"]
metainfo = metainfo[metainfo["group"].isin(labels)].reset_index(drop=True)
bulk_exp = pd.read_csv("/public/data/scDT/SingleCellDownload/IBD/Treatment_Bulk/Bulk_Anti-TNF/GSE16879/GSE16879.exp.csv",header=0,sep=',',index_col=0)
bulk_exp = bulk_exp.T
print (bulk_exp.shape)
bulk_genelist = bulk_exp.columns
sample_full_labels = sorted(list(set(metainfo["group"].tolist())))
sample_full_labels_dict = dict(zip(sample_full_labels,range(len(sample_full_labels)))) ## label对应的index
sample_dict = dict(zip(metainfo["SampleName"],metainfo["group"]))
#list_labels = []
#for x in range(len(sample_full_labels)):
#    list_labels.append([x])
#list_labels_tensor = torch.tensor(list_labels,dtype=torch.int64)
print (sample_full_labels_dict[sample_dict["GSM364647"]])
#print (list_labels_tensor)

sample_full_labels

sample_full_labels_dict

print (bulk_exp.iloc[0:5,0:5])
select_sample = metainfo[metainfo["group"].isin(labels)]["SampleName"]
bulk_exp_select = bulk_exp[bulk_exp.index.isin(select_sample)]
sample_index_dict = dict(zip(bulk_exp_select.index,range(bulk_exp_select.shape[0]))) ## 将id 同index 存为字典
print (bulk_exp_select.shape)

## 对于每一个reaction , PCA 得到的主成分 PCA1值
import glob
import re
import os
import math
from sklearn import preprocessing

## 对于每一个reaction , PCA 得到的主成分 PCA1值
from sklearn.decomposition import PCA
from sklearn.preprocessing import StandardScaler
pca = PCA(n_components=2)
print(datetime.datetime.now().strftime('%Y-%m-%d  %H:%M:%S'))
sample_reaction_pca = pd.DataFrame()
n = 0
reaction_PCA_values = []
for node in iter(uniq_nodes): ## 计算所有reaction节点 对应的PCA1 values
    #print (node)
    #node = "R-HSA-1006169"
    if (node in ReactionGene_dict.keys()):
        genelist = ReactionGene_dict[node].split(",")
        common_gene = list(set(genelist) & set(bulk_genelist))
        ## 去除bulk中不存在的基因
        sub_exp = bulk_exp_select.loc[:,common_gene]
        #print (sub_exp)
        if (sub_exp.shape[1] >= 2):
            #sub_exp = sub_exp.values
            sub_exp = StandardScaler().fit_transform(sub_exp.values)
            principalComponents = pca.fit_transform(sub_exp)
            PCA_values =  pd.DataFrame(data = principalComponents,columns = ["PCA1","PCA2"])
            node_PCA_values = PCA_values["PCA1"]
        elif (sub_exp.shape[1] == 1): ## 如果reaction中只包含一个基因，无法做PCA，则对该基因表达量进行标准化
            sub_exp = StandardScaler().fit_transform(sub_exp.values)
            node_PCA_values = pd.DataFrame(data = sub_exp, columns=["PCA1"])
            #print (node_PCA_values)
        else:
            node_PCA_values = pd.DataFrame(np.zeros((bulk_exp_select.shape[0],1)),columns=["PCA1"])
    else:
        node_PCA_values = pd.DataFrame(np.zeros((bulk_exp_select.shape[0],1)),columns=["PCA1"])
    reaction_PCA_values.append(node_PCA_values)
    #sample_reaction_pca = pd.concat([sample_reaction_pca,node_PCA_values],axis=1,join='outer',ignore_index=True)
    #n += 1
    #if n >3:break
    #break
sample_reaction_pca = pd.concat(reaction_PCA_values,axis=1,join='outer',ignore_index=True)
print(datetime.datetime.now().strftime('%Y-%m-%d  %H:%M:%S'))
print (sample_reaction_pca.iloc[0:10,0:10])
print (sample_reaction_pca.shape)


## 根据选取一半不同标签样本作为微调的训练集 
bulk_tuning_sets = []
bulk_test_sets = []
for group in labels:
    bulk_group_sample = metainfo[metainfo["group"] == group]["SampleName"]
    tuning_sample = bulk_group_sample.sample(math.ceil(bulk_group_sample.shape[0]/2)) ## 一半且向上取整
    print (tuning_sample)
    tuning_sample_index = tuning_sample.index
    print (group,len(tuning_sample))
    for m in tuning_sample:
        sampleName = m
        sample_index = sample_index_dict[sampleName]
        sample_value = sample_reaction_pca.loc[sample_index].values
        sample_value = sample_value.reshape(sample_value.shape[0],1)
        sample_node_feature = Data(x=torch.tensor(sample_value,dtype = torch.float32),y=torch.tensor(sample_full_labels_dict[sample_dict[sampleName]]),edge_index = edges)
        bulk_tuning_sets.append(sample_node_feature)
        
    test_sample = pd.concat([bulk_group_sample,tuning_sample,tuning_sample]).drop_duplicates(keep=False)
    for n in test_sample:
        sampleName = n
        sample_index = sample_index_dict[sampleName]
        sample_value = sample_reaction_pca.loc[sample_index].values
        sample_value = sample_value.reshape(sample_value.shape[0],1)
        sample_node_feature = Data(x=torch.tensor(sample_value,dtype = torch.float32),y=torch.tensor(sample_full_labels_dict[sample_dict[sampleName]]),edge_index = edges)
        bulk_test_sets.append(sample_node_feature)
        
print (len(bulk_tuning_sets))
print (len(bulk_test_sets))

print (sample_full_labels_dict)
print (sample_index_dict["GSM364656"])
#print (sample_index_dict["GSM422967"])
print (sample_full_labels_dict[sample_dict["GSM364656"]])


def train_tuning():
    model_tuning.train() ## 
    for data in bulk_total_loader:
        #print (data.y)
        data = data.to(device)
        #x = data.x.to(device)
        #e = data.edge_index.to(device)
        #b = data.batch.to(device)
        #y = data.y.to(device)
        optimizer_tuning.zero_grad()
        out = model_tuning(data.x, data.edge_index,data.batch)
        loss = criterion_tuning(out, data.y)
        loss.backward()
        optimizer_tuning.step()
        
        
def test_tuning(loader):
    model_tuning.eval()
    correct = 0
    for data in loader:                           # 批遍历测试集数据集。
        data = data.to(device)
        out = model_tuning(data.x, data.edge_index, data.batch) # 一次前向传播
        pred = out.argmax(dim=1)   # 使用概率最高的类别
        #print (pred,data.y)
        correct += int((pred == data.y).sum())           # 检查真实标签
    return correct / len(loader.dataset)


#model_tuning = GNN(num_node_features=1,hidden_channels=64,num_classes=31)
#checkpoint = torch.load('/public/home/liujunwu/workdir/scripts/GNN_Reactome/reaction_file/reaction.PCA.pth.tar')
#model_tuning.load_state_dict(checkpoint["model"])
#model_tuning.eval()
model_tuning = torch.load("/public/home/liujunwu/workdir/scripts/GNN_Reactome/reaction_file/reaction.PCA_convEdgeWeight.1018.pt",map_location=device)
model_tuning.lin = Linear(64,2)
print (model_tuning)
for name,layer in model_tuning.named_parameters():
    if (name == 'lin.weight' or (name == 'lin.bias')):
        #print (name,layer)
        layer.requires_grad = True
    else:
        layer.requires_grad = False
model_tuning = model_tuning.to(device)
#model_tuning

for name, param in model_tuning.named_parameters(): print(name, param)

model_tuning

#bulk_train_loader =  DataLoader(bulk_tuning_sets, batch_size=2,shuffle=True)
#bulk_test_loader = DataLoader(bulk_test_sets, batch_size=2,shuffle=True)
bulk_total_sets = bulk_tuning_sets + bulk_test_sets
bulk_total_loader = DataLoader(bulk_total_sets,batch_size=1,shuffle=False)
len(bulk_total_loader)

optimizer_tuning = torch.optim.AdamW(filter(lambda p: p.requires_grad, model_tuning.parameters()),lr=0.001)
criterion_tuning = torch.nn.CrossEntropyLoss()

print(datetime.datetime.now().strftime('%Y-%m-%d  %H:%M:%S'))

for epoch in range(1, 201):
    train_tuning()
    print(datetime.datetime.now().strftime('%Y-%m-%d  %H:%M:%S'))
    train_acc = test_tuning(bulk_total_loader)
    #test_acc = test_tuning(TCGA_test_loader)
    print(f'Epoch: {epoch:03d}, Train Acc: {train_acc:.4f}')
    if (train_acc > 0.85):
        break

for name, param in model_tuning.named_parameters(): print(name, param)

def test_tuning_subclass(loader):
    model_tuning.eval()
    correct = 0
    total = 0
    label_correct = {}
    total_label = {}
    for k in range(3):
        #print (type(k))
        total_label[k] = 0
        label_correct[k] = 0
    #print (label_correct[10])
    for data in loader:# 批遍历测试集数据集。
        data = data.to(device)
        out = model_tuning(data.x, data.edge_index, data.batch) # 一次前向传播
        pred = out.argmax(dim=1)   # 使用概率最高的类别
        total += int((pred == data.y).sum())
        index = 0
        for i in pred:
            num = i.item()
            #print (num)
            #print (data.y[0].item())
            total_label[data.y[index].item()] += 1
            if (num == data.y[index].item()):
                label_correct[num] += 1
            index += 1
    for j in total_label.keys():
        correct_num = label_correct[j]
        ratio = correct_num / total_label[j] if total_label[j] != 0 else 0
        print (j,total_label[j],ratio)
    return total / len(loader.dataset)

test_acc = test_tuning_subclass(bulk_test_loader)
#print (test_acc.)
print(f'Test Acc: {test_acc:.4f}')

from captum.attr import Saliency, IntegratedGradients

def model_forward(e_mask, dt):
    #batch = torch.zeros(dt.x.shape[0], dtype=int).to(device)
    batch = torch.zeros(dt.x.shape[0], dtype=int).to(device)
    out = model_tuning(dt.x,
                dt.edge_index,
                batch,
                e_mask)
    return out

def explain(method, dt, target=0):
    #input_mask = torch.ones(dt.edge_index.shape[1]).requires_grad_(True).to(device)
    input_mask = torch.ones(dt.edge_index.shape[1]).requires_grad_(True)
    if method == 'ig':
        ig = IntegratedGradients(model_forward)
        mask = ig.attribute(input_mask, target=target,
                            additional_forward_args=(dt,),
                            internal_batch_size=dt.edge_index.shape[1])
    elif method == 'saliency':
        saliency = Saliency(model_forward)
        mask = saliency.attribute(input_mask, target=target,
                                  additional_forward_args=(data,))
    else:
        raise Exception('Unknown explanation method')
 
    edge_mask = np.abs(mask.cuda().detach().numpy())
    if edge_mask.max() > 0:  # avoid division by zero
        edge_mask = edge_mask / edge_mask.max()
    return edge_mask

data = bulk_total_sets[10]
print (data)
data = data.to(device)
target = 0
target = torch.tensor(target).to(device)
#import random
#data = random.choice([t for t in bulk_test_sets if not t.y.item()])

target

torch.save(model_tuning, '/public/home/liujunwu/workdir/scripts/GNN_Reactome/reaction_file/GSE16879.IBD.Response.pt')

torch.save(bulk_total_sets,"/public/home/liujunwu/workdir/scripts/GNN_Reactome/reaction_file/GSE16879.IBD.Response.data.pt")

edge_mask = explain("ig", data, target=0)


for target_tissue in range(2):
    title = 'Integrated Gradients'
    method = 'ig'
    #data.to(device)
    #print(F"processing tissue {target_tissue} with {title}, a.k.a. {method}")
    edge_mask = explain(method, data, target=target_tissue)
    # edge_mask_dict = aggregate_edge_directions(edge_mask, data)
    #path = F"/home/jgburk/PycharmProjects/reticula/data/SRP035988/output/{method}_{target_tissue}.txt"
    #numpy.savetxt(path, edge_mask, delimiter=",")
    #print(F"{method} {target_tissue} edges saved as {path}")

#input_mask = torch.ones(data.edge_index.shape[1]).requires_grad_(True).to(device)
input_mask = torch.ones(data.edge_index.shape[1]).requires_grad_(True)
input_mask

ig = IntegratedGradients(model_forward)
ig.attribute

mask = ig.attribute(input_mask, target=1,
                            additional_forward_args=(data,),
                            internal_batch_size=data.edge_index.shape[1])
mask

import sklearn
t_v = np.loadtxt("/public/home/liujunwu/workdir/scripts/GNN_Reactome/reaction_file/test.label", dtype=str)
t_v
t_encoder = sklearn.preprocessing.LabelEncoder()
t_v = t_encoder.fit_transform(t_v)
torch.tensor(t_v[1])

data1 = DataLoader(bulk_total_sets)

print (data.edge_index.device)
print (data.x.device)
print (data.y.device)

data1.dataset[0].edge_index
